Move extent reading to route and not in DAO
