+++ app-backend/api/src/main/scala/scene/Routes.scala
@@ -24,6 +24,7 @@ import doobie.util.transactor.Transactor
 import com.azavea.rf.datamodel._
 import com.azavea.rf.database.filter.Filterables._
 import cats.implicits._
+import com.azavea.rf.common.utils.CogUtils
 import doobie._
 import doobie.implicits._
 import doobie.Fragments.in
@@ -91,7 +92,26 @@ trait SceneRoutes extends Authentication
   def createScene: Route = authenticate { user =>
     entity(as[Scene.Create]) { newScene =>
       authorize(user.isInRootOrSameOrganizationAs(newScene)) {
-        onSuccess(SceneDao.insert(newScene, user).transact(xa).unsafeToFuture) { scene =>
+
+        val tileFootprint = (newScene.sceneType, newScene.ingestLocation, newScene.tileFootprint) match {
+          case (Some(SceneType.COG), Some(ingestLocation), None) => {
+            logger.info(s"Generating Footprint for Newly Added COG")
+            CogUtils.getTiffExtent(ingestLocation)
+          }
+          case _ => {
+            logger.info("Not generating footprint, already exists")
+            None
+          }
+        }
+
+        val dataFootprint = (tileFootprint, newScene.dataFootprint) match {
+          case (Some(tf), None) => tileFootprint
+          case _ => newScene.dataFootprint
+        }
+
+        val updatedScene = newScene.copy(dataFootprint = dataFootprint, tileFootprint = tileFootprint)
+
+        onSuccess(SceneDao.insert(updatedScene, user).transact(xa).unsafeToFuture) { scene =>
           if (scene.statusFields.ingestStatus == IngestStatus.ToBeIngested) kickoffSceneIngest(scene.id)
           complete((StatusCodes.Created, scene))
         }
